this changes adds state tracing for all val object. This is required to ensure correctness with complex return objects. We should improve the performance future.
