+++ nautilus/include/nautilus/common/traceing.hpp
@@ -1,6 +1,7 @@
 
 #pragma once
 
+#include "nautilus/common/TypedValueRef.hpp"
 #include "nautilus/common/Types.hpp"
 #include <any>
 #include <array>
@@ -47,55 +48,8 @@ enum Op : uint8_t {
 	NEGATE,
 };
 
-using ValueRef = uint16_t;
 
-struct TypedValueRef {
-	ValueRef ref;
-	Type type;
 
-	TypedValueRef(uint16_t ref, Type type);
-	TypedValueRef();
-
-	bool operator==(const TypedValueRef& rhs) const {
-		return ref == rhs.ref;
-	}
-
-	bool operator!=(const TypedValueRef& rhs) const {
-		return !(rhs == *this);
-	}
-
-	bool operator<(const TypedValueRef& rhs) const {
-		return ref < rhs.ref;
-	}
-
-	bool operator>(const TypedValueRef& rhs) const {
-		return rhs < *this;
-	}
-
-	bool operator<=(const TypedValueRef& rhs) const {
-		return !(rhs < *this);
-	}
-
-	bool operator>=(const TypedValueRef& rhs) const {
-		return !(*this < rhs);
-	}
-
-	[[nodiscard]] std::string toString() const;
-};
-
-class TypedValueRefHolder {
-public:
-	TypedValueRefHolder(TypedValueRef valueRef);
-	TypedValueRefHolder(const TypedValueRefHolder&);
-	TypedValueRefHolder& operator=(const TypedValueRefHolder& other);
-	TypedValueRefHolder(TypedValueRefHolder&&);
-	TypedValueRefHolder& operator=(TypedValueRefHolder&& other);
-	~TypedValueRefHolder();
-	operator const TypedValueRef&() const;
-	TypedValueRef valueRef;
-};
-
-using value_ref = TypedValueRef;
 
 template <typename T>
 constexpr Type to_type();
