+++ nautilus/src/nautilus/CMakeLists.txt
@@ -1,4 +1,5 @@
 
+add_subdirectory(common)
 add_subdirectory(exceptions)
 add_subdirectory(tracing)
 add_subdirectory(compiler)
