+++ nautilus/src/nautilus/common/CMakeLists.txt
@@ -1,0 +1,2 @@
+ add_source_files(nautilus
+         TypedValueRefHolder.cpp)
